+++ front-end/src/app/shared/services/report-is-editable.service.spec.ts
