+++ front-end/src/app/transactions/transactions-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { ReportResolver } from 'app/shared/resolvers/report.resolver';
 import { TransactionTypeResolver } from 'app/shared/resolvers/transaction-type.resolver';
-import { ReportIsEditableGuard } from '../shared/guards/access-permissions.guards';
+import { ReportIsEditableGuard } from '../shared/guards/report-is-editable.guards';
 import { TransactionContainerComponent } from './transaction-container/transaction-container.component';
 import { TransactionTypePickerComponent } from './transaction-type-picker/transaction-type-picker.component';
 import { TransactionListComponent } from './transaction-list/transaction-list.component';
